Mise à jour

+++ src/Controller/CreationsController.php
@@ -21,14 +21,31 @@ class CreationsController extends AbstractController
     public function index(CreationsRepository $creationsRepository): Response
     {
         $creations = $creationsRepository->findAll();
+        
+        // On vérifie si l'utilisateur est connecté et a le rôle ROLE_USER
+        $isUser = $this->isGranted('ROLE_USER');
+        $isAdmin = $this->isGranted('ROLE_ADMIN');
 
         return $this->render('creations/index.html.twig', [
             'creations' => $creations,
+            'is_user' => $isUser,
+            'is_admin' => $isAdmin
+        ]);
+    }
+
+    #[Route('/{id}/personnaliser', name: 'app_personnaliser', methods: ['GET'])]
+    public function personnaliser(Creations $creation): Response
+    {
+        // Vérifie si l'utilisateur est connecté
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
+        return $this->render('creations/personnaliser.html.twig', [
+            'creation' => $creation,
         ]);
     }
 
     //Route menant vers la page de détail de chaques articles
-    #[Route('/creations/{id}', name: 'creation_details')]
+    #[Route('/details/{id}', name: 'creation_details')]
     public function details(EntityManagerInterface $entityManager, int $id): Response
     {
         $creation = $entityManager->getRepository(Creations::class)->find($id);
@@ -37,23 +54,21 @@ class CreationsController extends AbstractController
             throw $this->createNotFoundException('La création n\'existe pas');
         }
 
+        // On vérifie si l'utilisateur est connecté et a le rôle ROLE_USER
+        $isUser = $this->isGranted('ROLE_USER');
+
         return $this->render('creations/details.html.twig', [
             'creation' => $creation,
-        ]);
-    }
-
-    public function list(EntityManagerInterface $em): Response
-    {
-        $articles = $em->getRepository(Creations::class)->findAll();
-
-        return $this->render('creations/list.html.twig', [
-            'creations' => $creations,
+            'is_user' => $isUser
         ]);
     }
 
     #[Route('/new', name: 'app_creations_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        // Seuls les admins peuvent créer des articles
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $creation = new Creations();
         $form = $this->createForm(CreationsType::class, $creation);
         $form->handleRequest($request);
@@ -74,14 +89,20 @@ class CreationsController extends AbstractController
     #[Route('/{id}', name: 'app_creations_show', methods: ['GET'])]
     public function show(Creations $creation): Response
     {
+        $isUser = $this->isGranted('ROLE_USER');
+        
         return $this->render('creations/show.html.twig', [
             'creation' => $creation,
+            'is_user' => $isUser
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_creations_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Creations $creation, EntityManagerInterface $entityManager): Response
     {
+        // Seuls les admins peuvent éditer
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(CreationsType::class, $creation);
         $form->handleRequest($request);
 
@@ -97,19 +118,22 @@ class CreationsController extends AbstractController
         ]);
     }
 
-    #[Route('/creations/{id}/delete', name: 'app_delete_art', methods: ['POST'])]
-public function delete(Request $request, Creations $creation, EntityManagerInterface $entityManager, CsrfTokenManagerInterface $csrfTokenManager): Response
-{
-    $token = $request->request->get('_token');
-    if (!$csrfTokenManager->isTokenValid(new CsrfToken('delete'.$creation->getId(), $token))) {
-        throw new InvalidCsrfTokenException('Token CSRF invalide');
+    #[Route('/{id}/delete', name: 'app_delete_art', methods: ['POST'])]
+    public function delete(Request $request, Creations $creation, EntityManagerInterface $entityManager, CsrfTokenManagerInterface $csrfTokenManager): Response
+    {
+        // Seuls les admins peuvent supprimer
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+        $token = $request->request->get('_token');
+        if (!$csrfTokenManager->isTokenValid(new CsrfToken('delete'.$creation->getId(), $token))) {
+            throw new InvalidCsrfTokenException('Token CSRF invalide');
+        }
+
+        $entityManager->remove($creation);
+        $entityManager->flush();
+
+        $this->addFlash('success', 'La création a été supprimée avec succès.');
+
+        return $this->redirectToRoute('app_creations_index');
     }
-
-    $entityManager->remove($creation);
-    $entityManager->flush();
-
-    $this->addFlash('success', 'La création a été supprimée avec succès.');
-
-    return $this->redirectToRoute('app_creations');
-}
 }
